improved robustness of room and guest, remaining booking

+++ GuestExperience/Repositories/GuestRepository.cs
@@ -17,12 +17,16 @@ public class GuestRepository : IGuestRepository
     {
         try
         {
+            if (guest == null)
+            {
+                throw new RepositoryException("Guest data invalid");
+            }
             await _context.Guests.AddAsync(guest);
             await _context.SaveChangesAsync();
         }
         catch(System.Exception ex)
         {
-            throw new CreateGuestException("Error while adding guest", ex);
+            throw new RepositoryException($"Error while adding guest {ex.Message}");
         }
 
         return guest;
@@ -30,45 +34,64 @@ public class GuestRepository : IGuestRepository
 
   
 
-    public async Task<Guest> DeleteGuest(int guestId)
+    public async Task<bool> DeleteGuest(int guestId)
     {   
         
         try
         {
-            Guest guest = await _context.Guests.FindAsync(guestId) ?? throw new InvalidOperationException();
+
+            var guest = await _context.Guests.FindAsync(guestId);
+            if (guest == null)
+            {
+                return false;
+            }
+            
             _context.Guests.Remove(guest);
             await _context.SaveChangesAsync();
-            return guest;
+            return true;
         }
         catch (System.Exception e)
         {
-          throw new CreateGuestException("Error while deleting guest", e);
+          throw new RepositoryException($"Error while deleting guest{e.Message}");
         }
 
     }
 
     public async Task<Guest> GetGuestByIdAsync(int guestId)
     {
-        var guest = await _context.Guests.FindAsync(guestId) ?? throw new CreateGuestException("Error while getting guest");
+        
         try
         {
+            var guest = await _context.Guests.FindAsync(guestId);
+            if (guest == null)
+            {
+                throw new RepositoryException("Guest not found");
+            }
             return guest;
         }
-        catch(System.Exception e)
+        catch(System.Exception ex)
         {
-            throw new CreateGuestException("Error while getting guest", e);
+            throw new RepositoryException($"Error while getting guest{ex}");
         }
     }
 
-    public async Task<List<Guest>> GetGuests()
+   
+
+    public async Task<List<Guest>> GetAllGuestsAsync()
     {
         try
         {
-            return await _context.Guests.ToListAsync();
+            var guests = await _context.Guests.ToListAsync();
+            if (guests == null)
+            {
+                throw new RepositoryException("Unable to retrieve guest data");
+            }
+
+            return guests;
         }
-        catch
+        catch(System.Exception ex)
         {
-            throw new CreateGuestException("Error while getting guests");
+            throw new RepositoryException($"Error while getting guests{ex}");
         }
     }
 
@@ -76,13 +99,17 @@ public class GuestRepository : IGuestRepository
     {
         try
         {
+            if (guest == null)
+            {
+                throw new RepositoryException("Guest not found");
+            }
             _context.Guests.Update(guest);
             var something = await _context.SaveChangesAsync();
             return guest;
         }
         catch(System.Exception ex)
         {
-            throw new GuestServiceException("Error while updating guest", ex);
+            throw new RepositoryException($"Error while updating guest {ex.Message}");
         }
     }
 
@@ -90,16 +117,20 @@ public class GuestRepository : IGuestRepository
     {
         if (email == null)
         {
-            throw new BadHttpRequestException("no valid email provided");
+            throw new RepositoryException("no valid email provided");
         }
         try
         {
             var guest =  _context.Guests.FirstOrDefault(guest => guest.Email == email);
+            if (guest == null)
+            {
+                throw new RepositoryException($"No guest with email {email} found");
+            }
             return guest;
         }
-        catch (System.Exception e)
+        catch (System.Exception ex)
         {
-            throw new GuestServiceException("Error while getting guest", e);
+            throw new RepositoryException($"Error while getting guest {ex}");
         }
 
     }
